Fix db <-> model inconsistencies
* Make all DB primary keys of type `bigserial`
* Make foreign key references not-nullable
* Make corresponding changes to all domain models and their queries & repositories
* Add `timezone` field to `city`
* Fix generic query tests
* Add domain-specific query tests
* Updated todos

+++ src/main/scala/flightdatabase/api/endpoints/LanguageEndpoints.scala
@@ -28,7 +28,7 @@ class LanguageEndpoints[F[_]: Concurrent] private (prefix: String, algebra: Lang
     case req @ POST -> Root =>
       req
         .attemptAs[LanguageModel]
-        .foldF[ApiResult[Int]](
+        .foldF[ApiResult[Long]](
           _ => Applicative[F].pure(Left(EntryInvalidFormat)),
           algebra.createLanguage
         )
