+++ src/main/scala/flightdatabase/domain/airplane/AirplaneAlgebra.scala
@@ -5,9 +5,9 @@ import flightdatabase.domain.ApiResult
 trait AirplaneAlgebra[F[_]] {
   def getAirplanes: F[ApiResult[List[AirplaneModel]]]
   def getAirplanesOnlyNames: F[ApiResult[List[String]]]
-  def getAirplane(id: Int): F[ApiResult[AirplaneModel]]
+  def getAirplane(id: Long): F[ApiResult[AirplaneModel]]
   def getAirplanesByManufacturer(manufacturer: String): F[ApiResult[List[AirplaneModel]]]
-  def createAirplane(airplane: AirplaneModel): F[ApiResult[Int]]
+  def createAirplane(airplane: AirplaneModel): F[ApiResult[Long]]
   def updateAirplane(airplane: AirplaneModel): F[ApiResult[AirplaneModel]]
-  def removeAirplane(id: Int): F[ApiResult[Unit]]
+  def removeAirplane(id: Long): F[ApiResult[Unit]]
 }
