+++ src/main/scala/flightdatabase/domain/airport/AirportAlgebra.scala
@@ -5,12 +5,12 @@ import flightdatabase.domain.ApiResult
 trait AirportAlgebra[F[_]] {
   def getAirports: F[ApiResult[List[AirportModel]]]
   def getAirportsOnlyNames: F[ApiResult[List[String]]]
-  def getAirport(id: Int): F[ApiResult[AirportModel]]
+  def getAirport(id: Long): F[ApiResult[AirportModel]]
   def getAirportByIata(iata: String): F[ApiResult[AirportModel]]
   def getAirportByIcao(icao: String): F[ApiResult[AirportModel]]
   def getAirportByCity(city: String): F[ApiResult[List[AirportModel]]]
   def getAirportByCountry(country: String): F[ApiResult[List[AirportModel]]]
-  def createAirport(airport: AirportModel): F[ApiResult[Int]]
+  def createAirport(airport: AirportModel): F[ApiResult[Long]]
   def updateAirport(airport: AirportModel): F[ApiResult[AirportModel]]
-  def removeAirport(id: Int): F[ApiResult[Unit]]
+  def removeAirport(id: Long): F[ApiResult[Unit]]
 }
