+++ src/main/scala/flightdatabase/domain/city/CityAlgebra.scala
@@ -5,9 +5,9 @@ import flightdatabase.domain.ApiResult
 trait CityAlgebra[F[_]] {
   def getCities: F[ApiResult[List[CityModel]]]
   def getCitiesOnlyNames: F[ApiResult[List[String]]]
-  def getCity(id: Int): F[ApiResult[CityModel]]
+  def getCity(id: Long): F[ApiResult[CityModel]]
   def getCitiesByCountry(country: String): F[ApiResult[List[CityModel]]]
-  def createCity(city: CityModel): F[ApiResult[Int]]
+  def createCity(city: CityModel): F[ApiResult[Long]]
   def updateCity(city: CityModel): F[ApiResult[CityModel]]
-  def removeCity(id: Int): F[ApiResult[Unit]]
+  def removeCity(id: Long): F[ApiResult[Unit]]
 }
