+++ src/main/scala/flightdatabase/domain/currency/CurrencyAlgebra.scala
@@ -5,8 +5,8 @@ import flightdatabase.domain.ApiResult
 trait CurrencyAlgebra[F[_]] {
   def getCurrencies: F[ApiResult[List[CurrencyModel]]]
   def getCurrenciesOnlyNames: F[ApiResult[List[String]]]
-  def getCurrency(id: Int): F[ApiResult[CurrencyModel]]
-  def createCurrency(currency: CurrencyModel): F[ApiResult[Int]]
+  def getCurrency(id: Long): F[ApiResult[CurrencyModel]]
+  def createCurrency(currency: CurrencyModel): F[ApiResult[Long]]
   def updateCurrency(currency: CurrencyModel): F[ApiResult[CurrencyModel]]
-  def removeCurrency(id: Int): F[ApiResult[Unit]]
+  def removeCurrency(id: Long): F[ApiResult[Unit]]
 }
